input polling, event fixes, frame count, ptr fixes

+++ windmill/src/interface/window/event/mouse_position_event.cpp
@@ -7,9 +7,9 @@ namespace wm {
 		return key;
 	}
 
-	mouse_position_event::mouse_position_event(const ptr<wm::window> window, const glm::vec2 position): window_event(window), position(position) { }
+	mouse_position_event::mouse_position_event(const ptr<wm::window> window, const glm::dvec2 position): window_event(window), position(position) { }
 
-	glm::vec2 mouse_position_event::get_new_position() const {
+	glm::dvec2 mouse_position_event::get_new_position() const {
 		return position;
 	}
