+++ windmill/src/interface/window/event/mouse_position_event.h
@@ -7,12 +7,12 @@ namespace wm {
 
 	class WM_PUBLIC mouse_position_event: public window_event {
 	private:
-		glm::vec2 position;
+		glm::dvec2 position;
 	public:
 		static key<mouse_position_event> get_key();
 
-		mouse_position_event(const ptr<wm::window> window, const glm::vec2 position);
-		glm::vec2 get_new_position() const;
+		mouse_position_event(const ptr<wm::window> window, const glm::dvec2 position);
+		glm::dvec2 get_new_position() const;
 	};
 
 }
